Move env test to same test class
This avoids duplicate fixtures
